Catch Broken Pipe error
A Broken Pipe happens when the browser is quit by the user or crashes before the request/response sequence has finished, mostly happens when db queue is high and responses are taking a long time. by setting the max queue size back to the default 100 should also help prevent this.